+++ docs/source/release-notes.rst
@@ -1,6 +1,13 @@
 Release Notes
 =============
 
+Version 0.19.2
+______________
+- Plugins:
+
+  - apps-client: support `UNSECURED` authentication strategy.
+
+
 Version 0.19.1
 ______________
 - Plugins: